+++ .github/instructions/.dev.instructions.md
@@ -7,7 +7,9 @@ Provide project context and coding guidelines that AI should follow when generat
 
 ## 📚 Codebase Awareness
 - Before making any code changes, **read the current codebase carefully**, especially relevant components and modules.
-- Refer to all **living documents** in `/docs`, `README.md`, or `NOTES.md` to understand ongoing plans and system architecture.
+- Read **all documents inside the `/docs` directory** (such as `docs/dev_notes.md`, `docs/tasks.md`, `docs/api.md`, etc.) to understand the system, project goals, and what's in progress.
+- These documents are considered **living sources of truth** and must be consulted and updated accordingly.
+- Also refer to `README.md`, `NOTES.md`, and any inline comments for additional context.
 - Do not assume anything is unused or safe to remove unless clearly marked.
 
 ## ✨ Code Quality Guidelines
@@ -20,7 +22,7 @@ Act as a **senior developer**. Your output must:
 
 ## 🧾 Documentation Updates
 When adding or modifying features:
-- Update the relevant living document such as `docs/dev_notes.md`, `docs/tasks.md`, or shared Notion pages.
+- Update the relevant documents inside `/docs/` (e.g., `dev_notes.md`, `tasks.md`, `api.md`) or shared Notion pages.
 - Include:
   - ✅ Completed tasks
   - 🔧 Changes/fixes
@@ -62,7 +64,7 @@ To run or test API endpoints:
      python main.py
      ```
 
-> 🧠 AI should check that the backend is running properly before calling or suggesting API endpoints. If unsure, refer to the API docs or existing route definitions in `main.py` or related Flask Blueprint modules.
+> 🧠 AI should check that the backend is running properly before calling or suggesting API endpoints. If unsure, refer to route definitions in `main.py` or any Flask Blueprint modules.
 
 ## 🤝 Collaboration Rules
 
@@ -72,4 +74,4 @@ To run or test API endpoints:
 
 ## ✅ Summary Prompt for Copilot:
 
-> Act as a senior developer. Carefully read the code and docs before making changes. Use clean, readable code with DRY and SoC principles. Comment anything unclear. Update `docs/dev_notes.md` with ✅ tasks, 🔄 future plans, and 🔧 changes. Be beginner-friendly and respectful of existing logic. Use `uv sync` and activate the Python environment before running backend with `uv run python main.py`.
+> Act as a senior developer. Carefully read the code and all documents inside `/docs/` before making changes. Use clean, readable code with DRY and SoC principles. Comment anything unclear. Update `/docs/` files with ✅ tasks, 🔄 future plans, and 🔧 changes. Be beginner-friendly and respectful of existing logic. Use `uv sync` and activate the Python environment before running backend with `uv run python main.py`.
