+++ docs/dev_notes.md
@@ -2,8 +2,69 @@
 
 ## Recent Completed Tasks
 
-### ✅ Base64 Profile Image Implementation (Latest Update)
-**Date**: Current Session
+### ✅ Swagger UI Modular Refactoring (Latest Update)
+**Date**: July 24, 2025
+**Changes Made**:
+- 🔧 Refactored monolithic `swagger_ui.py` (1,530 lines) into modular structure
+- 🔧 Created `server_config.py` for environment and server configuration (162 lines)
+- 🔧 Created `schemas.py` for all OpenAPI component schemas (462 lines)
+- 🔧 Created `paths.py` for API endpoint path definitions (480 lines)
+- 🔧 Simplified `swagger_ui.py` to main UI rendering logic (187 lines)
+- 🔧 Added ReDoc alternative documentation interface
+- 🔧 Enhanced admin endpoint visibility in Swagger UI
+- 🔧 Improved styling and user experience
+
+**Modular Architecture**:
+```
+app/swagger/
+├── swagger_ui.py         # 187 lines - Main UI & routing
+├── server_config.py      # 162 lines - Config & metadata
+├── schemas.py           # 462 lines - Data models
+├── paths.py             # 480 lines - API endpoints
+└── swagger_ui_backup.py  # 1,530 lines - Original backup
+```
+
+**Benefits**:
+- ✅ **88% reduction** in main file size (1,530 → 187 lines)
+- ✅ **Clear separation** of concerns and responsibilities
+- ✅ **Easy maintenance** - developers can work on specific components
+- ✅ **Scalable architecture** - simple to add new endpoints/schemas
+- ✅ **Enhanced UX** - added ReDoc interface and improved styling
+- ✅ **Better organization** - logical grouping of related functionality
+
+**New Documentation Interfaces**:
+- **Swagger UI**: `http://localhost:5000/docs/` (enhanced styling)
+- **ReDoc**: `http://localhost:5000/docs/redoc` (clean alternative)
+- **Health Check**: `http://localhost:5000/docs/health` (service status)
+- **JSON Spec**: `http://localhost:5000/docs/swagger.json` (OpenAPI spec)
+
+### ✅ Admin Endpoints Integration (Previous Update)
+**Date**: July 24, 2025
+**Changes Made**:
+- 🔧 Fixed admin endpoints not appearing in Swagger UI
+- 🔧 Resolved namespace conflicts between actual and placeholder routes
+- 🔧 Added comprehensive admin endpoint documentation
+- 🔧 Updated admin tag description to remove "Coming Soon"
+
+**Admin Endpoints Available**:
+- **POST** `/api/admin/users/promote` - Promote user to admin
+- **GET** `/api/admin/users` - List all admin users
+- **GET** `/api/admin/users/{admin_id}` - Get specific admin details
+
+### ✅ Database-Aware Health Checks (Previous Update)
+**Date**: July 24, 2025
+**Changes Made**:
+- 🔧 Enhanced health check to detect database type automatically
+- 🔧 Added `get_database_info()` helper function
+- 🔧 Environment-specific queries: SQLite for local, PostgreSQL for production
+- 🔧 Added support for MySQL and fallback for unknown databases
+
+**Health Check Endpoints**:
+- **GET** `/api/health` - Basic health and connectivity check
+- **GET** `/api/health/detailed` - Detailed system information with database version
+
+### ✅ Base64 Profile Image Implementation (Completed)
+**Date**: Previous Session
 **Changes Made**:
 - 🔧 Added `profile_image` field to User model (Text type for Base64 storage)
 - 🔧 Created comprehensive image validation utility (`app/utils/image_utils.py`)
