+++ docs/project_status.md
@@ -1,15 +1,15 @@
 # WeRent Backend - Project Status
 
-**Last Updated:** July 22, 2025  
-**Project Phase:** Authentication Complete + Enhanced Documentation + Field Alignment Fixed (Item 'status' field fully removed and schemas aligned)  
+**Last Updated:** July 24, 2025  
+**Project Phase:** Authentication Complete + Admin Management + Modular Documentation Architecture  
 **Flask App Status:** ✅ Running on http://localhost:5000  
-**Documentation Status:** ✅ Interactive API docs available at /docs/ and /redoc/
+**Documentation Status:** ✅ Enhanced interactive API docs with modular architecture at /docs/ and /redoc/
 
 ---
 
 ## 📋 Project Overview
 
-**WeRent** is an equipment rental platform backend built with Flask. The project follows RESTful API design principles, uses JWT authentication for secure user management, and implements Flask best practices with a modular architecture. Features comprehensive interactive API documentation with OpenAPI 3.0 and enhanced Swagger UI.
+**WeRent** is an equipment rental platform backend built with Flask. The project follows RESTful API design principles, uses JWT authentication for secure user management, and implements Flask best practices with a modular architecture. Features comprehensive interactive API documentation with OpenAPI 3.0, enhanced Swagger UI, and modular documentation architecture for better maintainability.
 
 ### Tech Stack
 - **Framework:** Flask (Python)
@@ -17,8 +17,8 @@
 - **Authentication:** JWT with Flask-JWT-Extended
 - **Password Security:** Flask-Bcrypt
 - **Database ORM:** SQLAlchemy with Flask-Migrate
-- **Data Validation:** Pydantic v2 with advanced validators (schemas now match models; no unused fields like 'status')
-- **API Documentation:** OpenAPI 3.0 with Swagger UI, ReDoc
+- **Data Validation:** Pydantic v2 with advanced validators
+- **API Documentation:** OpenAPI 3.0 with modular Swagger UI, ReDoc
 - **Package Management:** UV (modern Python package manager)
 - **Environment:** Python Virtual Environment (.venv)
 - **Testing:** Pytest with Flask-Testing
@@ -27,6 +27,49 @@
 ---
 
 ## ✅ Completed Features
+
+### 🛡️ Admin Management System (NEW - COMPLETE)
+- [x] Admin User Promotion Endpoint
+- [x] List All Admin Users Endpoint  
+- [x] Get Specific Admin Details Endpoint
+- [x] Admin-only Route Protection with @admin_required decorator
+- [x] Comprehensive Admin API Documentation
+- [x] Admin Endpoints Integration in Swagger UI
+
+**Admin API Endpoints:**
+```
+POST /api/admin/users/promote    - Promote user to admin (Admin only)
+GET  /api/admin/users           - List all admin users (Admin only)
+GET  /api/admin/users/{id}      - Get admin user details (Admin only)
+```
+
+### 📊 Enhanced Documentation System (NEW - COMPLETE)
+- [x] Modular Swagger UI Architecture (88% file size reduction)
+- [x] Multiple Documentation Interfaces (Swagger UI + ReDoc)
+- [x] Enhanced Styling and User Experience
+- [x] Organized Schema and Path Management
+- [x] Environment-Aware Configuration
+- [x] Professional Documentation Presentation
+
+**Documentation Interfaces:**
+```
+GET  /docs/                     - Enhanced Swagger UI interface
+GET  /docs/redoc               - Alternative ReDoc interface  
+GET  /docs/swagger.json        - OpenAPI 3.0 specification
+GET  /docs/health              - Documentation service health
+```
+
+### 🏥 Database-Aware Health Monitoring (NEW - COMPLETE)
+- [x] Environment-Specific Database Queries
+- [x] Automatic Database Type Detection
+- [x] Comprehensive Health Check Endpoints
+- [x] Production vs Development Query Optimization
+
+**Health Check Endpoints:**
+```
+GET  /api/health               - Basic connectivity and status
+GET  /api/health/detailed      - Comprehensive system information
+```
 
 ### 🔐 Authentication System (COMPLETE)
 - [x] User Registration/Signup with validation (no JWT session created)
